requirements.txt
absl-py @ file:///croot/absl-py_1714140470852/work
aiohttp @ file:///croot/aiohttp_1715108769648/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
astunparse==1.6.3
attrs @ file:///croot/attrs_1695717823297/work
blinker @ file:///croot/blinker_1696539051175/work
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools @ file:///croot/cachetools_1713977085444/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1714483155441/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///croot/click_1698129812380/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy @ file:///croot/contourpy_1700583582875/work
cryptography @ file:///croot/cryptography_1694444244250/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
filelock @ file:///croot/filelock_1700591183607/work
flatbuffers @ file:///home/ktietz/cip/python-flatbuffers_1634039120618/work
fonttools @ file:///croot/fonttools_1713551344105/work
frozenlist @ file:///croot/frozenlist_1698702560391/work
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
gmpy2 @ file:///work/ci_py311/gmpy2_1676839849213/work
google-auth @ file:///croot/google-auth_1715111149956/work
google-auth-oauthlib @ file:///work/ci_py311_2/google-auth-oauthlib_1679340681059/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///croot/grpc-suite_1681912592597/work
h5py @ file:///croot/h5py_1715094721489/work
idna @ file:///croot/idna_1714398848350/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1710971335535/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1708996548741/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1717182742060/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1710257359434/work
keras @ file:///home/builder/mesters/opt/envs/tensorflow/conda-bld/keras_1682445665871/work/keras-2.12.0-py2.py3-none-any.whl
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver @ file:///work/ci_py311/kiwisolver_1676827230232/work
Markdown @ file:///work/ci_py311/markdown_1676837083112/work
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib @ file:///croot/matplotlib-suite_1713336378214/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict @ file:///croot/multidict_1701096859099/work
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682952636358/work
oauthlib @ file:///croot/oauthlib_1679489621486/work
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.2.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow @ file:///croot/pillow_1714398848491/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
ply==3.11
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==3.20.3
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1705722403006/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
PyJWT @ file:///croot/pyjwt_1715094751094/work
pynvml==11.5.0
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing @ file:///work/ci_py311/pyparsing_1677811559502/work
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq @ file:///croot/pyzmq_1705605076900/work
requests @ file:///croot/requests_1716902831423/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scipy @ file:///croot/scipy_1714069776893/work/dist/scipy-1.13.0-cp311-cp311-linux_x86_64.whl#sha256=5ea9bb3db7de954ca26bfbeb18592fabbe71154f6bc567a9ad09df45dd38f54d
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy @ file:///croot/sympy_1701397643339/work
tensorboard @ file:///home/builder/mesters/opt/envs/tensorflow/conda-bld/tensorboard_1682445826165/work/tensorboard-2.12.1-py3-none-any.whl
tensorboard-data-server @ file:///croot/tensorboard-data-server_1681498183723/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.6.0
tensorflow @ file:///home/builder/mesters/opt/envs/tensorflow/conda-bld/tensorflow-base_1682961422577/work/tensorflow_pkg/tensorflow-2.12.0-cp311-cp311-linux_x86_64.whl
tensorflow-estimator @ file:///home/builder/mesters/opt/envs/tensorflow/conda-bld/tensorflow-estimator_1682445976941/work/tensorflow_estimator-2.12.0-py2.py3-none-any.whl
termcolor @ file:///work/ci_py311/termcolor_1676838860478/work
torch==2.3.0
torch-tb-profiler==0.4.3
torchaudio==2.3.0
torchsummary==1.5.1
torchvision==0.18.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1708363099148/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
triton==2.3.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
urllib3 @ file:///croot/urllib3_1715635851070/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug @ file:///croot/werkzeug_1716533313107/work
wrapt @ file:///work/ci_py311/wrapt_1676827586170/work
yarl @ file:///croot/yarl_1701105127787/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
